+++ src/components/POLineCard/POLineCard.js
@@ -20,6 +20,7 @@ const propTypes = {
   poLine: PropTypes.shape({
     acquisitionMethod: PropTypes.string,
     id: PropTypes.string,
+    instanceId: PropTypes.any.string,
     poLineNumber: PropTypes.string,
     titleOrPackage: PropTypes.string,
   }).isRequired,
@@ -62,6 +63,16 @@ const POLineCard = ({
               {poLine.titleOrPackage ?? <NoValue />}
             </div>
           </KeyValue>
+          {poLine?.instanceId &&
+            <AppIcon app="inventory" iconKey="instance" size="small">
+              <Link
+                data-test-po-line-view-in-inventory
+                to={urls.viewInstance(poLine.instanceId)}
+              >
+                <FormattedMessage id="ui-agreements.poLines.viewInInventory" />
+              </Link>
+            </AppIcon>
+          }
         </Col>
       </Row>
     </div>